+++ src/Core/Plugin/PluginManager.php
@@ -20,15 +20,19 @@ use Symfony\Component\EventDispatcher\EventDispatcher;
  */
 class PluginManager
 {
-    private $plugins = [];
+    private $pluginCollection;
     private $eventSubscriberPlugins = [];
     private $eventDispatcher;
 
+    /**
+     * Constructor.
+     * 
+     * @param EventDispatcher $eventDispatcher The event dispatcher.
+     */
     public function __construct(EventDispatcher $eventDispatcher)
     {
         $this->eventDispatcher = $eventDispatcher;
-
-        $this->clearPlugin();
+        $this->pluginCollection = new PluginCollection();
     }
 
     /**
@@ -36,7 +40,7 @@ class PluginManager
      */
     public function callInitialize()
     {
-        foreach ($this->plugins as $plugin) {
+        foreach ($this->pluginCollection as $plugin) {
             $subscriber = new EventSubscriber();
             $plugin->initialize($subscriber);
 
@@ -47,6 +51,16 @@ class PluginManager
     }
 
     /**
+     * Gets the plugin collection.
+     * 
+     * @return Yosymfony\Spress\Core\Plugin\PluginCollection The plugin collection.
+     */
+    public function getPluginCollection()
+    {
+        return $this->pluginCollection;
+    }
+
+    /**
      * Releases resources like event listeners.
      */
     public function tearDown()
@@ -54,99 +68,6 @@ class PluginManager
         foreach ($this->eventSubscriberPlugins as list($plugin, $eventSubscriber)) {
             $this->removeListeners($plugin, $eventSubscriber);
         }
-    }
-
-    /**
-     * Adds a new plugin.
-     *
-     * @param string                                       $name   The plugin identifier.
-     * @param Yosymfony\Spress\Core\Plugin\PluginInterface $plugin The plugin.
-     */
-    public function addPlugin($name, PluginInterface $plugin)
-    {
-        if ($this->hasPlugin($name) === false) {
-            $this->setPlugin($name, $plugin);
-        }
-    }
-
-    /**
-     * Gets a plugin.
-     *
-     * @param string $name The plugin identifier.
-     *
-     * @return Yosymfony\Spress\Core\Plugin\PluginInterface
-     *
-     * @throws RuntimeException If the plugin is not defined.
-     */
-    public function getPlugin($name)
-    {
-        if ($this->hasPlugin($name) === false) {
-            throw new \RuntimeException(sprintf('Plugin "%s" not found.', $name));
-        }
-
-        return $this->plugins[$name];
-    }
-
-    /**
-     * Gets the plugins registered.
-     *
-     * @return Yosymfony\Spress\Core\Plugin\PluginInterface[] A key-value array with the name of the plugin as key.
-     */
-    public function getPlugins()
-    {
-        return $this->plugins;
-    }
-
-    /**
-     * Checks if a plugin exists.
-     *
-     * @param string $name The plugin identifier.
-     *
-     * @return bool
-     */
-    public function hasPlugin($name)
-    {
-        return isset($this->plugins[$name]);
-    }
-
-    /**
-     * Sets a plugin.
-     *
-     * @param string                                       $name   The name of the plugin.
-     * @param Yosymfony\Spress\Core\Plugin\PluginInterface $plugin The plugin.
-     */
-    public function setPlugin($name, PluginInterface $plugin)
-    {
-        $this->plugins[$name] = $plugin;
-    }
-
-    /**
-     * Counts the plugins registered.
-     *
-     * @return int
-     */
-    public function countPlugins()
-    {
-        return count($this->plugins);
-    }
-
-    /**
-     * Clears all generators registered.
-     */
-    public function clearPlugin()
-    {
-        $this->plugins = [];
-        $this->eventSubscriberPlugins = [];
-    }
-
-    /**
-     * Removes a plugin.
-     *
-     * @param string $name The plugin identifier.
-     */
-    public function removePlugin($name)
-    {
-        unset($this->plugins[$name]);
     }
 
     private function addListeners(PluginInterface $plugin, EventSubscriber $subscriber)